fix: change playground model config to metadata
