+++ packages/frameworks/vue/src/chat/index.ts
@@ -1,5 +1,5 @@
 export { default as GenuiConfigProvider } from './ConfigProvider.vue';
-export { default as GenuiChat } from './index.vue';
+export { default as GenuiChat } from './GenuiChat.vue';
 export * from './chat.types.js';
 export * from './injection-tokens.js';
 export * from './i18n/index.js';
